refactoring last status health check endpoint

+++ hermes/internal/notification/payloads/messagepayloads.go
@@ -54,6 +54,6 @@ type FullMessageResponse struct {
 }
 
 type StatusResponse struct {
-	Status  int16  `json:"status"`
+	Status  int    `json:"status"`
 	Details string `json:"details"`
 }
